% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clothing-data.R
\docType{data}
\name{clothing}
\alias{clothing}
\title{Cloth sales in China}
\format{
An object of class \code{"data.frame"}.
}
\usage{
data(clothing)
}
\description{
Daily sales, in natural logarithms, of a clothing brand in 25 provinces in China from January 1, 2008, to December 9, 2012. The number of observations is 1812.
}
\references{
Chang, J., Guo, B., and Yao, Q. (2018). Principal component analysis for second-order stationary vector time series. \emph{The Annals of Statistics}, 46(5), 2094-2124.
}
\keyword{datasets}
