% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPIEurope200015-data.R
\docType{data}
\name{CPIEurope200015}
\alias{CPIEurope200015}
\title{Price Indexes EUUS}
\format{
An object of class \code{"data.frame"}.
}
\source{
\href{https://ec.europa.eu/eurostat}{https://ec.europa.eu/eurostat}
}
\usage{
data("CPIEurope200015")
}
\description{
Monthly consumer price indexes of European countries and the United States of America
for the period January 2000 to October
2015. There are 33 indexes and the names of the countries are
the columns names. The number of observations is 190.
}
\keyword{datasets}
