% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FREDMDApril19.R
\docType{data}
\name{FREDMDApril19}
\alias{FREDMDApril19}
\title{Federal Reserve Bank at St Louis.}
\format{
An object of class \code{"data.frame"}.
}
\source{
\href{https://www.stlouisfed.org/}{https://www.stlouisfed.org/}
}
\usage{
data("FREDMDApril19")
}
\description{
Data obtained from the Federal Research Bank after process
to remove missing values.
}
\keyword{datasets}
