% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdpsimple6c8018-data.R
\docType{data}
\name{gdpsimple6c8018}
\alias{gdpsimple6c8018}
\title{Growth of GDP in 6 Countries}
\format{
An object of class \code{"data.frame"}.
}
\source{
\href{https://fred.stlouisfed.org/}{https://fred.stlouisfed.org/}
}
\usage{
data(gdpsimple6c8018)
}
\description{
\eqn{([x(t)-x(t-1)]/x(t-1))} of United States, United Kingdom, France, Australia, Germany, and Canada.
Quarterly data from 1980 to 2018. The original data are downloaded from FRED
(Federal Reserve Bank of St Louis). The GDP is based on expenditures.
}
\keyword{datasets}
