% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations032017-data.R
\docType{data}
\name{locations032017}
\alias{locations032017}
\title{Locations of Air-Box Devices in Taiwan}
\format{
An object of class \code{"data.frame"}.
}
\usage{
data(locations032017)
}
\description{
Latitude and longitude of the Air Boxes used in TaiwanAirBox032017.
}
\references{
Chen, L.J. et al. (2017). Open framework for participatory PM2.5 monitoring in smart cities. \emph{IEEE Access Journal},
Vol. 5, pp. 14441-14454.
}
\keyword{datasets}
