% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdec1to5-data.R
\docType{data}
\name{mdec1to5}
\alias{mdec1to5}
\title{Monthly Simple Returns of United States Market Portfolios.}
\format{
An object of class \code{"data.frame"}.
}
\source{
\href{https://www.crsp.org/}{https://www.crsp.org/}
}
\usage{
data("mdec1to5")
}
\description{
Based on Market Cap. The first 5 (smallest 10 percentage, next 10 percentage, etc.).
The time span is from January 1962 to December 2018. The data is from CRSP
(Center of Research for the Security Prices). The first column is calendar time.
}
\keyword{datasets}
