% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mexpimpcnus-data.R
\docType{data}
\name{mexpimpcnus}
\alias{mexpimpcnus}
\title{Monthly Exports and Imports of China and United States.}
\format{
An object of class \code{"data.frame"}.
}
\source{
\href{https://fred.stlouisfed.org/}{https://fred.stlouisfed.org/}
}
\usage{
data("mexpimpcnus")
}
\description{
The original data are from FRED (Federal Reserve Bank of St Louis) and the unit is US Dollars.
The time span is from January 1992 to December 2018.
}
\keyword{datasets}
