% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantilebox.R
\name{quantileBox}
\alias{quantileBox}
\title{Quantile Boxplot}
\usage{
quantileBox(x, prob = c(0.25, 0.5, 0.75))
}
\arguments{
\item{x}{T by k data matrix: T data points in rows with each row being data at a given time point,
and k time series in columns.}

\item{prob}{Probability, the quantile series of which is to be computed.
Default values are 0.25, 0.5, 0.75.}
}
\value{
Boxplot.
}
\description{
Boxplots of selected quantiles of each time series (Column-wise operations).
}
\examples{
data(TaiwanAirBox032017)
quantileBox(TaiwanAirBox032017[,1:3])

}
