% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stockindexes99world-data.R
\docType{data}
\name{Stockindexes99world}
\alias{Stockindexes99world}
\title{World Stock Indexes}
\format{
An object of class \code{"data.frame"}.
}
\usage{
data(Stockindexes99world)
}
\description{
Standardized daily stock indexes of the 99 most important financial markets
around the world from January 3, 2000, to December 16, 2015, with a total of
4163 observations. The first column contains the dates and the names of the indexes are
the columns names.
}
\references{
Refence or source
}
\keyword{datasets}
