% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaiwanAirBox032017-data.R
\docType{data}
\name{TaiwanAirBox032017}
\alias{TaiwanAirBox032017}
\title{Hourly PM25 Measurements from Air-Box Devices in Taiwan}
\format{
An object of class \code{"data.frame"}.
}
\source{
\href{https://sites.google.com/site/cclljj/research/dataset-airbox}{https://sites.google.com/site/cclljj/research/dataset-airbox}
}
\usage{
data(TaiwanAirBox032017)
}
\description{
Hourly PM25 measurements were constructed from random minute observations collected by AirBox devices for March 2017.
There are 744 observations and 516 series.
}
\references{
Chen, L.J. et al. (2017). Open framework for participatory PM2.5 monitoring in smart cities. \emph{IEEE Access Journal},
Vol. 5, pp. 14441-14454.
}
\keyword{datasets}
