% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperatures-data.R
\docType{data}
\name{temperatures}
\alias{temperatures}
\title{World Temperatures}
\format{
An object of class \code{"data.frame"}.
}
\source{
\href{https://www.ncdc.noaa.gov/cag/}{https://www.ncdc.noaa.gov/cag/}
}
\usage{
data(temperatures)
}
\description{
Three series with 106 observations (from year 1910 to 2016) with the deviation with respect to average value
of temperatures in November in three regions: Europe, North America and South America. First columns contains the years.
Units are degrees Celsius.
}
\keyword{datasets}
