% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMEdata20002018-data.R
\docType{data}
\name{UMEdata20002018}
\alias{UMEdata20002018}
\title{Quarterly Economic Series of the European Monetary Union}
\format{
An object of class \code{"data.frame"}.
}
\source{
\href{https://ec.europa.eu/eurostat}{https://ec.europa.eu/eurostat}
}
\usage{
data("UMEdata20002018")
}
\description{
Series of Gross Domestic Product at market prices, Household and NPISH final consumption expenditure and Gross
Fixed Capital Formation for the 19 Euro Area country members, a total of 57 series.
The source of the data is Eurostat and the data was extracted 08-07-2019.
Seasonally and calendar adjusted data. The data includes 76 quaterly observations from Q1-2000 to Q4-2018.
}
\keyword{datasets}
